adding rest client support
