+++ src/main/kotlin/io/github/ydwk/ydwk/impl/YDWKImpl.kt
@@ -25,21 +25,9 @@ import io.github.ydwk.ydwk.YDWK
 import io.github.ydwk.ydwk.cache.*
 import io.github.ydwk.ydwk.entities.Application
 import io.github.ydwk.ydwk.entities.Bot
-import io.github.ydwk.ydwk.entities.Guild
-import io.github.ydwk.ydwk.entities.User
 import io.github.ydwk.ydwk.entities.application.PartialApplication
 import io.github.ydwk.ydwk.entities.channel.DmChannel
-import io.github.ydwk.ydwk.entities.channel.TextChannel
-import io.github.ydwk.ydwk.entities.channel.VoiceChannel
-import io.github.ydwk.ydwk.entities.channel.guild.Category
-import io.github.ydwk.ydwk.entities.guild.Member
 import io.github.ydwk.ydwk.entities.message.embed.builder.EmbedBuilder
-import io.github.ydwk.ydwk.event.backend.event.CoroutineEventListener
-import io.github.ydwk.ydwk.event.backend.event.GenericEvent
-import io.github.ydwk.ydwk.event.backend.event.IEventListener
-import io.github.ydwk.ydwk.event.backend.managers.CoroutineEventManager
-import io.github.ydwk.ydwk.event.backend.managers.SampleEventManager
-import io.github.ydwk.ydwk.impl.entities.UserImpl
 import io.github.ydwk.ydwk.impl.entities.channel.DmChannelImpl
 import io.github.ydwk.ydwk.impl.entities.message.embed.builder.EmbedBuilderImpl
 import io.github.ydwk.ydwk.impl.rest.RestApiManagerImpl
@@ -47,31 +35,33 @@ import io.github.ydwk.ydwk.impl.slash.SlashBuilderImpl
 import io.github.ydwk.ydwk.rest.EndPoint
 import io.github.ydwk.ydwk.rest.RestApiManager
 import io.github.ydwk.ydwk.slash.SlashBuilder
-import io.github.ydwk.ydwk.ws.WebSocketManager
-import io.github.ydwk.ydwk.ws.util.GateWayIntent
 import io.github.ydwk.ydwk.ws.util.LoggedIn
-import java.time.Instant
 import java.util.*
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.TimeUnit
-import kotlin.random.Random
 import okhttp3.OkHttpClient
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
-class YDWKImpl(
+open class YDWKImpl(
     private val client: OkHttpClient,
-    private val simpleEventManager: SampleEventManager = SampleEventManager(),
-    private val coroutineEventManager: CoroutineEventManager = CoroutineEventManager(),
     val logger: Logger = LoggerFactory.getLogger(YDWKImpl::class.java),
-    private val allowedCache: MutableSet<CacheIds> = mutableSetOf(),
-    val cache: Cache = PerpetualCache(allowedCache),
-    val memberCache: MemberCache = MemberCacheImpl(allowedCache),
-    private var token: String? = null,
+    var token: String? = null,
     private var guildIdList: MutableList<String> = mutableListOf(),
     var applicationId: String? = null
 ) : YDWK {
-    // logger
+
+    final override var loggedInStatus: LoggedIn? = null
+        private set
+
+    /**
+     * Sets the logged in status
+     *
+     * @param loggedIn The logged in status which is used to send messages to discord.
+     */
+    fun setLoggedIn(loggedIn: LoggedIn) {
+        this.loggedInStatus = loggedIn
+    }
 
     override val objectNode: ObjectNode
         get() = JsonNodeFactory.instance.objectNode()
@@ -79,23 +69,7 @@ class YDWKImpl(
     override val objectMapper: ObjectMapper
         get() = ObjectMapper()
 
-    override var webSocketManager: WebSocketManager? = null
-        private set
-
-    override fun shutdownAPI() {
-        val oneToFiveSecondTimeout = Random.nextLong(1000, 5000)
-        invalidateRestApi(oneToFiveSecondTimeout)
-        logger.info("Timeout for $oneToFiveSecondTimeout then shutting down")
-
-        try {
-            Thread.sleep(oneToFiveSecondTimeout)
-        } catch (e: InterruptedException) {
-            logger.error("Error while sleeping", e)
-        }
-        webSocketManager?.shutdown()
-    }
-
-    private fun invalidateRestApi(oneToFiveSecondTimeout: Long) {
+    fun invalidateRestApi(oneToFiveSecondTimeout: Long) {
         client.dispatcher.executorService.shutdown()
         client.connectionPool.evictAll()
         client.dispatcher.cancelAll()
@@ -110,21 +84,11 @@ class YDWKImpl(
             oneToFiveSecondTimeout, TimeUnit.MILLISECONDS)
     }
 
-    override fun getGuild(id: String): Guild? {
-        return cache[id, CacheIds.GUILD] as Guild?
-    }
-
-    override fun getGuilds(): List<Guild> {
-        return cache.values(CacheIds.GUILD).map { it as Guild }
-    }
-
     override val restApiManager: RestApiManager
         get() {
             val botToken = token ?: throw IllegalStateException("Bot token is not set")
             return RestApiManagerImpl(botToken, this, client)
         }
-    override val uptime: Instant
-        get() = webSocketManager!!.upTime ?: throw IllegalStateException("Bot is not logged in")
 
     override val slashBuilder: SlashBuilder
         get() =
@@ -137,40 +101,8 @@ class YDWKImpl(
         guildIds.forEach { this.guildIdList.add(it) }
     }
 
-    override fun setAllowedCache(vararg cacheTypes: CacheIds) {
-        allowedCache.addAll(cacheTypes.toSet())
-    }
-
-    override fun setDisallowedCache(vararg cacheTypes: CacheIds) {
-        allowedCache.removeAll(cacheTypes.toSet())
-    }
-
-    override fun getTextChannel(id: Long): TextChannel? {
-        return cache[id.toString(), CacheIds.TEXT_CHANNEL] as TextChannel?
-    }
-
-    override fun getTextChannels(): List<TextChannel> {
-        return cache.values(CacheIds.TEXT_CHANNEL).map { it as TextChannel }
-    }
-
-    override fun getVoiceChannel(id: Long): VoiceChannel? {
-        return cache[id.toString(), CacheIds.VOICE_CHANNEL] as VoiceChannel?
-    }
-
-    override fun getVoiceChannels(): List<VoiceChannel> {
-        return cache.values(CacheIds.VOICE_CHANNEL).map { it as VoiceChannel }
-    }
-
     override val embedBuilder: EmbedBuilder
         get() = EmbedBuilderImpl(this)
-
-    override fun getCategory(id: Long): Category? {
-        return cache[id.toString(), CacheIds.CATEGORY] as Category?
-    }
-
-    override fun getCategories(): List<Category> {
-        return cache.values(CacheIds.CATEGORY).map { it as Category }
-    }
 
     override fun createDmChannel(userId: Long): CompletableFuture<DmChannel> {
         return this.restApiManager
@@ -184,34 +116,6 @@ class YDWKImpl(
                     DmChannelImpl(this, jsonBody, jsonBody["id"].asLong())
                 }
             }
-    }
-
-    override fun getMember(guildId: Long, userId: Long): Member? {
-        return memberCache[userId.toString(), guildId.toString()] as Member?
-    }
-
-    override fun getMembers(): List<Member> {
-        return memberCache.values(CacheIds.MEMBER).map { it as Member }
-    }
-
-    override fun getUser(id: Long): User? {
-        return cache[id.toString(), CacheIds.USER] as User?
-    }
-
-    override fun getUsers(): List<User> {
-        return cache.values(CacheIds.USER).map { it as User }
-    }
-
-    override fun requestUser(id: Long): CompletableFuture<User> {
-        return this.restApiManager.get(EndPoint.UserEndpoint.GET_USER, id.toString()).execute { it
-            ->
-            val jsonBody = it.jsonBody
-            if (jsonBody == null) {
-                throw IllegalStateException("json body is null")
-            } else {
-                UserImpl(jsonBody, jsonBody["id"].asLong(), this)
-            }
-        }
     }
 
     override var bot: Bot? = null
@@ -257,128 +161,6 @@ class YDWKImpl(
             } // wait for application to be set
             return field
         }
-
-    override var loggedInStatus: LoggedIn? = null
-        private set
-
-    @get:Synchronized
-    override val waitForConnection: YDWK
-        get() {
-            val ws = webSocketManager
-            if (ws == null) {
-                throw IllegalStateException("WebSocketManager is not initialized")
-            } else {
-                while (!ws.connected) {
-                    try {
-                        Thread.sleep(100)
-                    } catch (e: InterruptedException) {
-                        logger.error("Error while waiting for connection", e)
-                    } finally {
-                        if (ws.connected) {
-                            logger.info("WebSocketManager connected")
-                        } else {
-                            logger.info("WebSocketManager not connected, retrying")
-                            waitForConnection // retry
-                        }
-                    }
-                }
-            }
-            return this
-        }
-    override val waitForReady: YDWK
-        get() {
-            val ws = webSocketManager
-            if (ws == null) {
-                throw IllegalStateException("WebSocketManager is not initialized")
-            } else {
-                while (!ws.ready) {
-                    try {
-                        Thread.sleep(100)
-                    } catch (e: InterruptedException) {
-                        logger.error("Error while waiting for ready", e)
-                    } finally {
-                        if (!ws.ready) {
-                            waitForReady // retry
-                        }
-                    }
-                }
-            }
-            return this
-        }
-
-    override fun addEvent(vararg eventListeners: Any) {
-        for (eventListener in eventListeners) {
-            when (eventListener) {
-                is IEventListener -> {
-                    simpleEventManager.addEvent(eventListener)
-                }
-                is CoroutineEventListener -> {
-                    coroutineEventManager.addEvent(eventListener)
-                }
-                else -> {
-                    logger.error(
-                        "Event listener is not an instance of EventListener or CoroutineEventListener")
-                }
-            }
-        }
-    }
-
-    override fun removeEvent(vararg eventListeners: Any) {
-        for (eventListener in eventListeners) {
-            when (eventListener) {
-                is IEventListener -> {
-                    simpleEventManager.removeEvent(eventListener)
-                }
-                is CoroutineEventListener -> {
-                    coroutineEventManager.removeEvent(eventListener)
-                }
-                else -> {
-                    logger.error(
-                        "Event listener is not an instance of EventListener or CoroutineEventListener")
-                }
-            }
-        }
-    }
-
-    override fun emitEvent(event: GenericEvent) {
-        simpleEventManager.emitEvent(event)
-        coroutineEventManager.emitEvent(event)
-    }
-
-    /**
-     * Starts the websocket manager
-     *
-     * @param token The token of the bot which is used to authenticate the bot.
-     * @param intents The gateway intent which will decide what events are sent by discord.
-     */
-    fun setWebSocketManager(token: String, intents: List<GateWayIntent>) {
-        var ws: WebSocketManager? = null
-        ws = WebSocketManager(this, token, intents)
-        this.webSocketManager = ws.connect()
-        this.timer(Timer(), ws)
-        this.token = token
-    }
-
-    @Synchronized
-    private fun timer(timer: Timer, ws: WebSocketManager) {
-        timer.scheduleAtFixedRate(
-            object : TimerTask() {
-                override fun run() {
-                    ws.sendHeartbeat()
-                }
-            },
-            0,
-            14 * 24 * 60 * 60 * 1000)
-    }
-
-    /**
-     * Sets the logged in status
-     *
-     * @param loggedIn The logged in status which is used to send messages to discord.
-     */
-    fun setLoggedIn(loggedIn: LoggedIn) {
-        this.loggedInStatus = loggedIn
-    }
 
     companion object {
         val logger: Logger = LoggerFactory.getLogger(YDWKImpl::class.java)
